// src/components/Search/reducer/enums/searchActions.enum.ts
export enum SearchActions {
  SET_PLACE = "SET_PLACE",
  SET_CURRENT_RECOMMENDED_PLACE = "SET_CURRENT_RECOMMENDED_PLACE",
  SET_HAS_RECOMMENDED_PLACE_CHANGED = "SET_HAS_RECOMMENDED_PLACE_CHANGED",
  SET_FLIGHT_TYPE = "SET_FLIGHT_TYPE",
  SET_FLIGHT_SETTINGS = "SET_FLIGHT_SETTINGS",
  SET_CABIN_CLASS = "SET_CABIN_CLASS",
  SET_PASSENGERS = "SET_PASSENGERS",
  SET_SEARCH_FORM_DATA = "SET_SEARCH_FORM_DATA",
  SET_LIMIT = "SET_LIMIT",
  SET_IS_LOADING = "SET_IS_LOADING",
  SET_SHOW_RESULTS = "SET_SHOW_RESULTS",
  SET_SEARCH_RESULTS = "SET_SEARCH_RESULTS",
  SET_INPUT_TEXTS = "SET_INPUT_TEXTS",
  SET_IS_ERROR = "SET_IS_ERROR",
  SET_VISIBLE_ITEMS = "SET_VISIBLE_ITEMS",
}